requirements.txt
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
confusable-homoglyphs==3.2.0
cryptography==2.7
defusedxml==0.6.0
Django==2.2.1
django-allauth==0.39.1
django-appconf==1.0.3
django-crispy-forms==1.7.2
django-extensions==2.1.9
idna==2.8
oauthlib==3.0.1
Pillow==6.0.0

pycparser==2.19
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
transliterate==1.10.2
urllib3==1.25.3
Werkzeug==0.15.4

